+++ src/CslaGeneratorSerialization/Extensions/IServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using Csla.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace CslaGeneratorSerialization.Extensions;
 
@@ -6,7 +7,13 @@ public static class IServiceCollectionExtensions
 {
 	public static IServiceCollection AddCslaGeneratorSerialization(this IServiceCollection self)
 	{
-		self.AddSingleton<CustomSerializationResolver>();
+		// We register MobileFormatterOptions because,
+		// if we need to use MobileFormatter to handle serialization
+		// for IMobileObject-based objects that
+		// do not participate in generator serialization,
+		// this type is needed by MobileFormatter.
+		self.AddSingleton<CustomSerializationResolver>()
+			.AddSingleton<MobileFormatterOptions>();
 		return self;
 	}
 }